Added bg20 and bgDemographicSubgroups2015to2019 etc
bgDemographicSubgroups2015to2019 will be for bg21
bgDemographicSubgroups2014to2018 is for bg20
bg20 is now included (EJSCREEN data released in Jan 2021, called the 2020 version of EJSCREEN, using ACS from 2014-2018.
Added a lot of detailed documentation of bg20 and especially of the bgDemographicSubgroups2014to2018 dataset that has percent hispanic, etc. that are not in EJSCREEN csv on FTP site.