Clarified documentation of bg20 and bgDemographicSubgroups2014to2018
Clarified documentation of bg20 and bgDemographicSubgroups2014to2018  noting FIPS and lowinc count and other errors in a very small number of rows, and how bg20 has PR but other does not right now.  Might fix those here and or may get fixed in EJSCREEN dataset that is on EPA's FTP site.